move(client): class, constants 파일 이동
각 class, constants 파일들이 점점 추가될 것으로 예상되어, component 폴더의 하위 폴더에 있는 것이 아닌 각 목적에 맞는 폴더에 추가될 수 있도록 두 파일을 이동함.
